Refactor MethodAndRawURICallback out of GatewayBridge
